Chapter 04/identity-auth-examples/requirements.txt
bcrypt==4.1.2
pyotp==2.9.0
PyJWT==2.8.0
cryptography>=41.0.0